Create Print row wise with condition.java
Once upon a time, there was a programmer named Alex who was given the task of printing a matrix row-wise. However, there was a twist - the even-numbered rows had to be printed from left to right, and the odd-numbered rows had to be printed from right to left.

help Alex and write a program that would iterate through each row of the matrix and check if it was an even or odd row. If it was an even row, the program would traverse it from left to right, and if it was an odd row, the program would traverse it from right to left.

Input Format

First line contains, m and n depicting the size of first matrix.

Second line contains m * n Integer values, depicting all the elements of matrix.