Create HW_Constructing a spiral Matrix.java
John is a software engineer assigned with the responsibility of developing a matrix visualization tool for a significant scientific dataset. The dataset is in the form of a one-dimensional integer array containing N × M elements, and John's task is to display the information in a spiral pattern that commences from the top-left corner and advances inwards. To accomplish this, John must create an N × M matrix that follows the specified spiral sequence from the input array. Would you be able to assist John with his project?

Input Format

first line contains integer k which represents the no. of elments of matrix. second and third line of input contains integer n and m which represents the row and column of tha matrix. fourth line of input contains k integers.

Constraints

1 <= k <= 10000

1 <= n , m <= 100

0 <= matrix[i][j] <= 10000

Output Format

print the matrix.

Sample Input 0

36 6 6
182 224 745 908 978 965 841 265 942 129 359 562 775 118 875 646 19 722 25 54 918 791 466 692 793 481 568 36 15 283 393 912 529 47 222 193
Sample Output 0

182 224 745 908 978 965 
54 918 791 466 692 841 
25 912 529 47 793 265 
722 393 193 222 481 942 
19 283 15 36 568 129 
646 875 118 775 562 359